[FIX] website,web_editor: allow to re-edit social media elements
This commit allows you to change the title and the icons of custom links
of the social media block after the first save. If an element is in a
parent that has the class `o_not_editable` then it is only editable if
it has the contenteditable attribute set to true. We rightfully remove
this attribute at saving but without putting it back when we return to
edit mode. This commit allows to fix this case. Note that we have to
improve this system in the future (so we can get rid of this patch).

This part of the fix is very similar to [this other fix] for the team
block.

Note that to allow the edition of custom icons, it was necessary to
patch the sanitize of web_editor because this one added
`contenteditable="false"` on all the icons which have the `fa` class.

[this other fix]: https://github.com/odoo/odoo/commit/b7d53d2fcb20a447fd3098adf94ffb33bc49dda9

task-3073171