[IMP] core: switch to native namespace packages
Seems to work fine as-is, with one nit: natively
`odoo.addons.__path__` is a custom sort-of live pseudo-collection, it
can be appended to and checked for containing but that's about it.

However we have a few concatenations (`+=`) and a `remove`, so we need
a list. Since we've always had it as a list that probably is not an
issue.