[FIX] website: allow to re-edit image comparison snippet elements
This commit allows you to change the images and the captions of the
image comparison block after the first save. If an element is in a
parent that has the class `o_not_editable` then it is only editable if
it has the contenteditable attribute set to true. We rightfully remove
this attribute at saving but without putting it back when we return to
edit mode. This commit allows to fix this case. Note that we have to
improve this system in the future (so we can get rid of this patch).

task-2531109