[IMP] mail: make composer mixin take lang from template
Mail composer mixin already synchronizes subject and body based on a mail
template, easing calling rendering method available through mail render
mixin.

In this commit we add the same mechanism on lang. It means that lang is
taken from the template and can be used to computed based on language
defined on template, generally based on a customer.

Task-