[IMP] web: adapt draggable_hook_builder for new sortableList hook
Purpose:
--------
Knowledge's sidebar revamp introduces a new draggable hook "SortableList"
that is meant to replace the usage of jQuery's sortable widget and of the
NestedSortable library previously used for the sidebar.

To be able to build this hook using the `draggable_hook_builder`, some
changes have been introduced:

- Add a `delay` parameter which is used to delay the start of the dragging
sequence. It is useful when movable elements are also clickable, to make
sure the mousedown event was triggered because the user wanted to
move an element, rather than simply click on it. The dragging sequence will
not be initiated if a mouseup event was triggered before `delay` ms after
the mousedown.

- Do not consider children elements of the currentElement as eligible
targets (to prevent nesting an element using one of its child as new parent)

- Return the value of the built hook handlers (SortableList has a handler
used to check if the position of the dragged element is valid)

- Do not remove the height of the dragged element when computing its max
allowed height (so that the dragged element is below the cursor even when
the cursor is at the max allowed height, instead of being on top of the
lasts elements, obstructing the view)

- Fix the computation of the container's height, which did not take into
account the case where the container was not at the top of its scrollable
container, which did not trigger the scroll when the cursor was at the
bottom of the container

- Allow to define mandatory params in the built hook's default params, so
that one do not have to define this mandatory param for each instance of
the hook (eg. SortableList is intended to only work with `li` elements)

Task-3074123