[FIX] web: misalignment title of dialog on Mobile
The title is aligned to the right on mobile. This is the result of the
two rules of Bootstrap:
```css
.modal-header {
  display: flex;
  justify-content: space-between;
}
```

To avoid this we add `margin-inline: auto`, to force center alignment of
the title.

Steps to reproduce:
* Open Sales (on small/mobile screen)
* Create a new Quotation
* Click on Customer field => BUG title of dialog on the right

Note: on desktop there is no problem as there is the close button on
the right.