[IMP] payment{_*}:add extra states for providers
Some providers require additional allowed states due to their
refund or transaction process justifying it.
Until now, these extra states were specified in payment module,
which was not ideal as it allowed every provider in every flow to
accept these additional states.

With this commit, additional states are now specified only in
the coresponding flow of a provider that requires them.

task-286967