[IMP] onboarding: support toggle per company
With many tests, keep support for changing onboarding & steps `is_per_company`.

As onboarding and their steps `is_per_company` must always match, and as
it is not possible (simply) to update records simultaneously, changing
these values is only possible when records are not linked or via a new
custom action.

This is why the fields are `readonly` on the form views when onboarding and
steps are linked, and why a check is added in the `write` methods as well.

Task-3025136