[IMP] website_sale, purchase: fix condition on portal_customer existence
Actually the group should always be present because it is added by portal.mixin
inherited in the 2 cases by sale.order but in this task, we encountered a case
where this inheritance wasn't working revealing that problem (condition of the
"if" cannot be false because the generator will generate an exception before).
So it doesn't fix an actual problem but a potential one.

Task-2674649