[IMP] crm: time per stages
This commit introduces a new setting in crm: time in stage.
This adds a button in the status bar: users can see how much time a record has
spent in each stage in the status bar.

This commit introduces:
	- the new setting in CRM
	- a new JS field and template that extends the statusbar field and adds the
	  new feature
	- a mixin, as a transferrable approach was requested. It adds a Char
	  computed field, which contains a JSON and can be parsed by the client. The
	  computation is based on the tracking messages of the record

All of this is accessible from the CRM module but can be extracted and used
independantly when needs be.

Task-3032773