[FIX] web: set input background to transparent
Prior to this commit, the `input-bg` in enterprise and in community were
inconsistent. Some inputs such as the ones in list view were appearing
in community with a weird white background while being hovered on.
Meanwhile in enterprise, these inputs were having a transparent
background.

This commit sets the `input-bg` variable in `bootstrap_overridden`
to transparent for the inputs in community. Since the variable is now
the same in community and enterprise, this variable is now unnecessary
and was removed from the enterprise branch.

Task-3094005

Enterprise PR: