[FIX] website,web_editor: prevent horizontal scroll on we-matrix
Before this commit, we-matrix could create a horizontal scroll bar in
the editor. This was because the inputs had a minimum width size.
Finally, a better solution is to distribute the available space between
the columns.
Steps to reproduce the bug:
- Drop a chart block on a page
- Add some series

=> The matrix overflows from the editor.

There is no perfect solution to this problem... We have the choice
between:
1) leave the existing overflow on the editor
2) put a horizontal scroll on the we-matrix
3) do the solution proposed in this commit.

This solution has a drawback which is that the cells can become really
small if there are a lot of columns but this solution seems to be the
least bad from a UX point of view.

task-3094162