[REF] mail: move composer 'record_name' field from onchange to computed
PURPOSE

Purpose of this task is to remove the _onchange_template_id method on composer
model. Split it into editable computed stored fields. It gives a better control
of value generation and avoid having to call the onchange when composer is
invoked in code.

SPECIFICATIONS
Improve composer to correctly compute field ``record_name`` in compute method
instead of in a semi-default-get method called, we hope, at the right time.

Computation is coming either from parent message, either being the display
name of the record when being in monorecord comment mode. In batch mode it
makes no sense to compute a single record name. In email mode it is not
used anyway. Computing it in composer allows to benefit from its computation
to give it to the message created afterwards during message_post.

Task-2088884 (Mail: Use editable computed stored fields in composer)