[IMP] website_event{_sale}: improve registered tag with visitor
When a visitor register to an event ticket which needs to be paid
if the visitor does not end the checkout he is still marked has
registered to the user.
We now display the registered tag info only when the registration
has been paid to not induce the user in error.

task-2992722