[FIX] web: update the type of orderBy props of the View component
Currently, there is a missmatch between the props definition of the `View`
component and the relational model:

For the `View` component, one can specify an `orderBy` props to order the
records based on a given key. The component expects the props to be an
array of strings (eg: `['phone ASC']`). However, the relational model
expects the `orderBy` props to be an array of JSON objects
(eg: `[{'name': 'phone', asc: true }]`).

When the `View` component passes the `orderBy` variable to the relational
model, owl raises an error because the relational model does not expect
the `orderBy` props to be an array of String.

This commit will update the `orderBy` props definition of the `View`
component and the `WithSearch` component to accept an array of JSON
objects. With that change, we will no longer get an error when setting
an order from the properties of the `View` component.

task-3056268