[IMP] account : total amount on invoice in letters
Some localizations require to print the total amount of an invoice in full letters. This commit enables this functionality. As this is optional, a boolean field in the accounting settings enables to activate the feature. The total amount is displayed at the bottom of any invoice if the company has the boolean fields in the accounting settings set to True (set by default to False)

task-3097097
PR-107714