[IMP] e-commerce : hiding non-available shipping methods while ordering
When the customer is ordering products, all the available shipping methods are displayed. If one of those is based on rule and that those rules are not respected, then a error message is displayed and the customer is not allowed to choose it. It is then unnecessary to display it.

task-3062417