[FIX] web_editor: prevent editing of header
## Issue

The way the live HTML editor for html_field works means the header
(and other nodes that should only appear once) are removed
and the content inserted inside the body.
This results in values being edited without the user being aware.

For emails, this results in any email using 'style' in the header
to break on GMail and other clients that do not support 'style'
outside of the header.

## Fix

Parse the value in html_field to check whether there are any head nodes.
If there are, reformat the value to pass it to the children components
and anywhere it is used for presentation.

If there was a node in <head>, re-use the original parsed document we
got from the value and just update the body  when updating the value.

## Assumption

This only fixes the issue for documents that required the <head>.
If a document needs to be stored in HTML but doesn't store anything in
<head>, it will still be trimmed after this fix.

task - 3035163