[IMP] web_editor: enable to use image options after saving an image
Before this commit, options like "Filter", "Width" and "Quality"
disappeared after replacing and saving a background image or any images
stored in the database. The goal of this commit is to enable the use of
these options even after saving a replaced image.

For background images, the problem was that their "src" attribute was
not correctly updated when initializing the image. This is due to the
fact that the target used to recover the url of the background image
was the snippet and not the image itself. This problem is resolved by
ensuring that the argument of the function "getBgImageURL" is the
background image. Now that the "src" attribute is correctly updated,
the "computeVisibility" function of the the "BackgroudOptimize" option
works properly and the options "Filter" "Width" and "Quality" are
displayed as wanted.

For the other images, the problem was solved in two steps:
1) Thanks to the resource model and id, the attachment related to the
image is provided to the "image processing" part. Thanks to it, the
function "loadImageInfo" fills the dataset of the image with the
original id, the original source and the mimetype of the image. Due to
that, the function "computeWidgetVisibility" of the Image tools option
can work properly and the options "Shape", "Filter", "Width" and
"Quality" are displayed as wanted.

2) The second part of the solution focuses on a way to save the data
related to the shape, the filter, the width and the quality inside the
database. Indeed, this is necessary in order to be able to display the
correct parameters at the re-edition. To do so, the choice has been
made to use the "description" field of the attachment to store these
information. Thanks to that, at the re-edition, the shape, filter,
width and quality data are updated in the DOM and correctly displayed.

Now that those options are always usable, an other problem appears.
Indeed, note that for products that do not have variants, changing the
image of the template will automatically change the image of the
product and vice versa. Due to that, a bug could occur:
- Change the image of the template.
- Go to the product and try to change the width of the product image.
- Bug: the old product image appears.
Indeed, in this case, a new attachment is created for the template but
no attachment is created for the product. When trying to modify the
width of the product image, the image on which the modifications apply
is still the old one. To solve this problem:
- A new attachment is created for the product each time its image
template is modified, only if the product does not have variants or if
the product has variants but does not have images. Indeed, in this last
case, the image of the template is used by default as the image of the
product.
- A new attachment is created for the template each time its product
image is modified, only if the product does not have variants.

task-3046425