[REF] onboarding: add safe close & set_done actions
This eases the handling of setting steps done or closing an
onboarding panel that may have been deleted. The code for
each panel/step can therefore be simplified.

This also allows to own in one place the default API/return
values for these actions and to avoid much duplication.

Task-3025136