[FIX] account: don't check reversed move state when reconciling bank statement
When reconciling a bank statement line with a lot of invoices (more than 400),
the process might timeout because it takes too much time checking if the
move state should be 'reversed'.

Since we are in a bank reconciliation, we can skip this check because
invoices should not be 'reversed' so this is what is done here.

Locally, when reconciling one statement line and 521 invoices:
- Without the fix: 7 min 17 s
- With the fix: 1 min 52 s

opw-3013391