[REF] mail: move composer 'auto_delete' field from onchange to computed
PURPOSE

Purpose of this task is to remove the _onchange_template_id method on composer
model. Split it into editable computed stored fields. It gives a better control
of value generation and avoid having to call the onchange when composer is
invoked in code.

SPECIFICATIONS

Improve composer to correctly compute field ``auto_delete`` in compute method
instead of being updated by an onchange and a default_get we have to call
manually.

Computation is

  * if a template is set, use its value;
  * otherwise, default is True for comment mode (remove notification emails)
    and False in email mode (keep complete archives by default);

Task-2088884 (Mail: Use editable computed stored fields in composer)