[IMP] website: prevent user from removing last block of snippet
The user has the option to remove the last block from the Table of Content
snippet prior to this commit, but doing so breaks the Add Item feature because
there is no reference to adding a new item in the snippet area, and doing so is
also not a good idea.
If the block or content is the final one in the Table of Content after this
commit, an alert message will appear and the user cannot remove it.

task-3081442