[FIX] tools, *: ignore the active field when updating assets
*: mass_mailing, web_editor, website, website_blog, website_mail_group,
   website_mass_mailing, website_payment, website_sale,
   test_assetsbundle, test_website

Before this commit when a module was updated all assets were reset to
their predefined `active` state.
This causes assets related to old snippet versions to be made inactive
even if those old snippet versions are used inside existing pages.

It used to work when the activation of assets was made through view
inheritance because when views are defined through a `<template>` tag,
the `active` attribute is intentionally ignored during updates
since [1], except for new records since [2].

This commit introduces an `<asset>` tag in the XML import format.
It is an alias of `<record ... model="ir.asset">` with the additional
feature that it avoids taking the `active` field into account during
updates for existing `ir_asset` records.
We then rely on the `website_disable_unused_snippets_assets` cron to
properly deactive any unused asset at a later stage.

Another approach was to overload `_load_records_write` in `base`'s
`ir_asset.py` to avoid taking the `active` field into account when
updating records:
```py
    def _load_records_write(self, values):
        values.pop('active', None)
        super()._load_records_write(values)
```
But this is not stable because it changes the way `ir.asset` records
are imported when the `<record>` tag is used.

[1]: https://github.com/odoo/odoo/commit/2d296cb77922d33be2dc45b900191fac34bda429#diff-175c28787c272a219b9275f79262a48af9aa029e718f45077fd609737559e84eR803-R804
[2]: https://github.com/odoo/odoo/commit/f1c70d4cc943ac4eb81a85a9dc005de34cd2060a#diff-175c28787c272a219b9275f79262a48af9aa029e718f45077fd609737559e84eR801-R804

task-2963840