-- For testing purpose only -- (to remove before r+)
This commit deactivates the observer at the drag start. This is done in
PR [1] but it was added in this PR in order to have the full fix in the
runbot so it can be tested easily by India. As mentioned, this commit
must be deleted before merging the PR (since this change will be
applied by [1]).

[1]: https://github.com/odoo/odoo/pull/106029