[FIX] survey: fix validation on conditional question
What are the steps to reproduce your issue?
Create a survey with single select multi choice question
Create a 2nd question which is also a single-select multi-choice question, which appears based on the answer selected for 1st question.
Create a 3rd question which is a text question which is visible based on the answer to question 2. All 3 questions are mandatory
Take the survey
Answer 1st question in a way the 2nd question appears
Answer 2nd question in a way 3rd one appears
Now change the answer the 1st question in a way 2nd question does not appear.
Submit the survey

What is the current behavior that you observe?
Odoo will not allow to submit the survey as it will require an answer for the 2nd question, even though it should not appear.

What would be your expected behavior in this case?
to be able to submit the survey.